Add JS file for functionality
